app/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
redis==4.5.5
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.52.0